ideater/modules/code_breakdown/requirements.txt
# Core dependencies
pydantic>=1.10.7
python-dotenv>=1.0.0

# OpenAI integration
openai>=0.27.4

# GitHub integration
pygithub>=1.58.1
gitpython>=3.1.31

# JIRA integration
jira>=3.5.0

# Code analysis and generation
ast>=0.0.2
astunparse>=1.6.3
black>=23.3.0

# Utilities
numpy>=1.24.3
tqdm>=4.65.0
python-slugify>=8.0.1
jinja2>=3.1.2